+++ src/inlay.ts
@@ -20,7 +20,7 @@ declare interface InlayHintEffect {
 const cls = 'cm-inlay-hint';
 
 class InlayHintWidget extends WidgetType {
-	declare label: string;
+	declare label;
 
 	constructor(label: string) {
 		super();