+++ src/matchBrackets.ts
@@ -217,8 +217,8 @@ export default (configs?: BracketConfig): Extension => {
 			}
 		},
 		{
-			decorations({decorations}) {
-				return decorations;
+			decorations(v) {
+				return v.decorations;
 			},
 		},
 	);